chore: add version request in snapshots
