+++ test/e2e/atlas/clusters/sharded/clusterssharded/clusters_sharded_test.go
@@ -53,7 +53,7 @@ func TestShardedCluster(t *testing.T) {
 	region, err := g.NewAvailableRegion(tier, e2eClusterProvider)
 	req.NoError(err)
 
-	mdbVersion, err := internal.MongoDBMajorVersion()
+	mdbVersion, err := g.MongoDBMajorVersion()
 	req.NoError(err)
 
 	g.Run("Create sharded cluster", func(t *testing.T) { //nolint:thelper // g.Run replaces t.Run
